Backend/requirements.txt
Django==3.2.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
gunicorn==20.1.0
djangorestframework==3.13.1
djoser==2.1.0
Pillow==9.0.0
django-cors-headers==3.11.0